+++ QueryType.cs
@@ -2,25 +2,25 @@ namespace mifty
 {
     public static class QueryType
     {
-        public const byte A = 1;
-        public const byte NS = 2;
-        public const byte MD = 3;
-        public const byte MF = 4;
-        public const byte CNAME = 5;
-        public const byte SOA = 6;
-        public const byte MB = 7;
-        public const byte MG = 8;
-        public const byte MR = 9;
-        public const byte NULL = 10;
-        public const byte WKS = 11;
-        public const byte PTR = 12;
-        public const byte HINFO = 13;
-        public const byte MINFO = 14;
-        public const byte MX = 15;
-        public const byte TXT = 16;
-        public const byte AFXR = 252;
-        public const byte MAILB = 253;
-        public const byte MAILA = 254;
-        public const byte All = 255;
+        public const ushort A = 1;
+        public const ushort NS = 2;
+        public const ushort MD = 3;
+        public const ushort MF = 4;
+        public const ushort CNAME = 5;
+        public const ushort SOA = 6;
+        public const ushort MB = 7;
+        public const ushort MG = 8;
+        public const ushort MR = 9;
+        public const ushort NULL = 10;
+        public const ushort WKS = 11;
+        public const ushort PTR = 12;
+        public const ushort HINFO = 13;
+        public const ushort MINFO = 14;
+        public const ushort MX = 15;
+        public const ushort TXT = 16;
+        public const ushort AFXR = 252;
+        public const ushort MAILB = 253;
+        public const ushort MAILA = 254;
+        public const ushort All = 255;
     }
 }
